Major refactoring: separated vector_model and evaluation related modules into multiple parts Added empty modules for clustering, to be filled later on
This refactoring aims to reach to the following goals:
1) It is a step towards extending sensai with models that are not based on vector models
2) There is generally more structure which makes reading and extending the library easier.
   The previously very large modules are broken into smaller separate chunks

There is only one minor breaking change (apart from the imports): the instantiation of evaluators
and cross-validators has been moved from a staticmethod away to a separate method in util
This was necessary in order to prevent circular imports.

While I have tried very hard to not add anything new and only to move existing stuff around,
I did fix some typos and added minor aesthetic changes along the way
