+++ docs/sensai/catboost.rst
@@ -1,0 +1,6 @@
+catboost
+========
+
+.. automodule:: sensai.catboost
+   :members:
+   :undoc-members:
