+++ docs/sensai/clustering/greedy_clustering.rst
@@ -1,0 +1,6 @@
+greedy\_clustering
+==================
+
+.. automodule:: sensai.clustering.greedy_clustering
+   :members:
+   :undoc-members:
