+++ docs/sensai/data_ingest.rst
@@ -1,0 +1,6 @@
+data\_ingest
+============
+
+.. automodule:: sensai.data_ingest
+   :members:
+   :undoc-members:
