+++ docs/sensai/evaluation/clustering.rst
@@ -1,0 +1,11 @@
+clustering
+==========
+
+.. automodule:: sensai.evaluation.clustering
+   :members:
+   :undoc-members:
+
+.. toctree::
+   :glob:
+
+   clustering/*
