+++ docs/sensai/evaluation/clustering/evaluators.rst
@@ -1,0 +1,6 @@
+evaluators
+==========
+
+.. automodule:: sensai.evaluation.clustering.evaluators
+   :members:
+   :undoc-members:
