+++ docs/sensai/evaluation/clustering/metrics.rst
@@ -1,0 +1,6 @@
+metrics
+=======
+
+.. automodule:: sensai.evaluation.clustering.metrics
+   :members:
+   :undoc-members:
