+++ docs/sensai/evaluation/eval_stats.rst
@@ -1,0 +1,6 @@
+eval\_stats
+===========
+
+.. automodule:: sensai.evaluation.eval_stats
+   :members:
+   :undoc-members:
