+++ docs/sensai/evaluation/evaluators.rst
@@ -1,0 +1,6 @@
+evaluators
+==========
+
+.. automodule:: sensai.evaluation.evaluators
+   :members:
+   :undoc-members:
