+++ docs/sensai/evaluation/util.rst
@@ -1,0 +1,6 @@
+util
+====
+
+.. automodule:: sensai.evaluation.util
+   :members:
+   :undoc-members:
