+++ docs/sensai/evaluation/vector_model.rst
@@ -1,0 +1,11 @@
+vector\_model
+=============
+
+.. automodule:: sensai.evaluation.vector_model
+   :members:
+   :undoc-members:
+
+.. toctree::
+   :glob:
+
+   vector_model/*
