+++ docs/sensai/evaluation/vector_model/eval_stats.rst
@@ -1,0 +1,11 @@
+eval\_stats
+===========
+
+.. automodule:: sensai.evaluation.vector_model.eval_stats
+   :members:
+   :undoc-members:
+
+.. toctree::
+   :glob:
+
+   eval_stats/*
