+++ docs/sensai/evaluation/vector_model/eval_stats/base.rst
@@ -1,0 +1,6 @@
+base
+====
+
+.. automodule:: sensai.evaluation.vector_model.eval_stats.base
+   :members:
+   :undoc-members:
