+++ docs/sensai/evaluation/vector_model/eval_stats/classification.rst
@@ -1,0 +1,6 @@
+classification
+==============
+
+.. automodule:: sensai.evaluation.vector_model.eval_stats.classification
+   :members:
+   :undoc-members:
