+++ docs/sensai/evaluation/vector_model/eval_stats/regression.rst
@@ -1,0 +1,6 @@
+regression
+==========
+
+.. automodule:: sensai.evaluation.vector_model.eval_stats.regression
+   :members:
+   :undoc-members:
