+++ docs/sensai/evaluation/vector_model/evaluators.rst
@@ -1,0 +1,11 @@
+evaluators
+==========
+
+.. automodule:: sensai.evaluation.vector_model.evaluators
+   :members:
+   :undoc-members:
+
+.. toctree::
+   :glob:
+
+   evaluators/*
