+++ docs/sensai/evaluation/vector_model/evaluators/base.rst
@@ -1,0 +1,6 @@
+base
+====
+
+.. automodule:: sensai.evaluation.vector_model.evaluators.base
+   :members:
+   :undoc-members:
