+++ docs/sensai/evaluation/vector_model/evaluators/classification.rst
@@ -1,0 +1,6 @@
+classification
+==============
+
+.. automodule:: sensai.evaluation.vector_model.evaluators.classification
+   :members:
+   :undoc-members:
