+++ docs/sensai/evaluation/vector_model/evaluators/regression.rst
@@ -1,0 +1,6 @@
+regression
+==========
+
+.. automodule:: sensai.evaluation.vector_model.evaluators.regression
+   :members:
+   :undoc-members:
