+++ docs/sensai/evaluation/vector_model/util.rst
@@ -1,0 +1,6 @@
+util
+====
+
+.. automodule:: sensai.evaluation.vector_model.util
+   :members:
+   :undoc-members:
