+++ docs/sensai/models.rst
@@ -1,0 +1,11 @@
+models
+======
+
+.. automodule:: sensai.models
+   :members:
+   :undoc-members:
+
+.. toctree::
+   :glob:
+
+   models/*
