+++ docs/sensai/models/base.rst
@@ -1,0 +1,6 @@
+base
+====
+
+.. automodule:: sensai.models.base
+   :members:
+   :undoc-members:
