+++ docs/sensai/util/hash.rst
@@ -1,0 +1,6 @@
+hash
+====
+
+.. automodule:: sensai.util.hash
+   :members:
+   :undoc-members:
