+++ src/sensai/__init__.py
@@ -1,19 +1,15 @@
-from .vector_model import VectorModel, VectorRegressionModel, VectorClassificationModel, InputOutputData, PredictorModel
-from .evaluation import VectorRegressionModelEvaluator, VectorRegressionModelEvaluationData, VectorRegressionModelCrossValidator, VectorRegressionModelCrossValidationData, \
-    VectorClassificationModelEvaluator, VectorClassificationModelEvaluationData, VectorClassificationModelCrossValidator, VectorClassificationModelCrossValidationData
-from .normalisation import NormalisationMode
-from . import eval_stats
-from . import sklearn
-from . import naive_bayes
-from . import util
-from . import hyperopt
-from .data_transformation import DataFrameTransformer, RuleBasedDataFrameTransformer
-from . import data_transformation
 from . import columngen
-from . import local_search
-from . import nearest_neighbors
+from . import data_transformation
 from . import featuregen
+from . import hyperopt
+from . import local_search
+from . import naive_bayes
+from . import nearest_neighbors
+from . import sklearn
+from . import util
+from .data_transformation import DataFrameTransformer, RuleBasedDataFrameTransformer
 from .ensemble import AveragingVectorRegressionModel
+from .normalisation import NormalisationMode
 
 # The following submodules are not imported by default to avoid necessarily requiring their dependencies:
 # tensorflow
