+++ src/sensai/evaluation/vector_model/__init__.py
@@ -1,0 +1,3 @@
+from .evaluators.base import VectorModelEvaluator, VectorModelCrossValidator
+from .evaluators.classification import VectorClassificationModelEvaluator, VectorClassificationModelCrossValidator
+from .evaluators.regression import VectorRegressionModelCrossValidator, VectorRegressionModelEvaluator
