+++ src/sensai/evaluation/vector_model/eval_stats/base.py
@@ -1,0 +1,121 @@
+from abc import ABC
+from typing import List, Dict, Generator, Tuple, Any, Union, TypeVar
+
+import numpy as np
+import pandas as pd
+
+from ...eval_stats import EvalStats, TMetric, ModelEvaluationData, EvalStatsCollection
+from ....models.vector_model import VectorModel
+from ....util.typing import PandasNamedTuple
+
+PredictionArray = Union[np.ndarray, pd.Series, pd.DataFrame, list]
+TEvalStats = TypeVar("TEvalStats", bound="VectorModelEvalStats")
+TVectorModelEvalStats = TypeVar("TVectorModelEvalStats", bound="VectorModelEvalStats")
+
+
+class VectorModelEvalStats(EvalStats[TMetric], ABC):
+    """Collects data for the evaluation of a model and computes corresponding metrics"""
+    def __init__(self, y_predicted: PredictionArray = None, y_true: PredictionArray = None,
+             metrics: List[TMetric] = None):
+        self.y_true = []
+        self.y_predicted = []
+        self.y_true_multidim = None
+        self.y_predicted_multidim = None
+        if y_predicted is not None:
+            self._addAll(y_predicted, y_true)
+        super().__init__(metrics)
+
+    def _add(self, y_predicted, y_true):
+        """
+        Adds a single pair of values to the evaluation
+
+        Parameters:
+            y_predicted: the value predicted by the model
+            y_true: the true value
+        """
+        self.y_true.append(y_true)
+        self.y_predicted.append(y_predicted)
+
+    def _addAll(self, y_predicted, y_true):
+        """
+        Adds multiple predicted values and the corresponding ground truth values to the evaluation
+
+        Parameters:
+            y_predicted: pandas.Series, array or list of predicted values or, in the case of multi-dimensional models,
+                        a pandas DataFrame (multiple series) containing predicted values
+            y_true: an object of the same type/shape as y_predicted containing the corresponding ground truth values
+        """
+        if (isinstance(y_predicted, pd.Series) or isinstance(y_predicted, list) or isinstance(y_predicted, np.ndarray)) \
+                and (isinstance(y_true, pd.Series) or isinstance(y_true, list) or isinstance(y_true, np.ndarray)):
+            a, b = len(y_predicted), len(y_true)
+            if a != b:
+                raise Exception(f"Lengths differ (predicted {a}, truth {b})")
+        elif isinstance(y_predicted, pd.DataFrame) and isinstance(y_true, pd.DataFrame):  # multiple time series
+            # keep track of multidimensional data (to be used later in getEvalStatsCollection)
+            y_predicted_multidim = y_predicted.values
+            y_true_multidim = y_true.values
+            dim = y_predicted_multidim.shape[1]
+            if dim != y_true_multidim.shape[1]:
+                raise Exception("Dimension mismatch")
+            if self.y_true_multidim is None:
+                self.y_predicted_multidim = [[] for _ in range(dim)]
+                self.y_true_multidim = [[] for _ in range(dim)]
+            if len(self.y_predicted_multidim) != dim:
+                raise Exception("Dimension mismatch")
+            for i in range(dim):
+                self.y_predicted_multidim[i].extend(y_predicted_multidim[:, i])
+                self.y_true_multidim[i].extend(y_true_multidim[:, i])
+            # convert to flat data for this stats object
+            y_predicted = y_predicted_multidim.reshape(-1)
+            y_true = y_true_multidim.reshape(-1)
+        else:
+            raise Exception(f"Unhandled data types: {str(type(y_predicted))}, {str(type(y_true))}")
+        self.y_true.extend(y_true)
+        self.y_predicted.extend(y_predicted)
+
+
+class VectorModelEvaluationData(ModelEvaluationData[TEvalStats], ABC):
+    def __init__(self, statsDict: Dict[str, TEvalStats], inputData: pd.DataFrame, model: VectorModel):
+        """
+        :param statsDict: a dictionary mapping from output variable name to the evaluation statistics object
+        :param inputData: the input data that was used to produce the results
+        :param model: the model that was used to produce predictions
+        """
+        self.evalStatsByVarName = statsDict
+        self.predictedVarNames = list(self.evalStatsByVarName.keys())
+        super().__init__(inputData, model)
+
+    def getEvalStats(self, predictedVarName=None) -> TEvalStats:
+        if predictedVarName is None:
+            if len(self.evalStatsByVarName) != 1:
+                raise Exception(f"Must provide name of predicted variable name, as multiple variables were predicted {list(self.evalStatsByVarName.keys())}")
+            else:
+                predictedVarName = next(iter(self.evalStatsByVarName.keys()))
+        evalStats = self.evalStatsByVarName.get(predictedVarName)
+        if evalStats is None:
+            raise ValueError(f"No evaluation data present for '{predictedVarName}'; known output variables: {list(self.evalStatsByVarName.keys())}")
+        return evalStats
+
+    def getDataFrame(self):
+        """
+        Returns an DataFrame with all evaluation metrics (one row per output variable)
+
+        :return: a DataFrame containing evaluation metrics
+        """
+        statsDicts = []
+        varNames = []
+        for predictedVarName, evalStats in self.evalStatsByVarName.items():
+            statsDicts.append(evalStats.getAll())
+            varNames.append(predictedVarName)
+        df = pd.DataFrame(statsDicts, index=varNames)
+        df.index.name = "predictedVar"
+        return df
+
+    def iterInputOutputGroundTruthTuples(self, predictedVarName=None) -> Generator[Tuple[PandasNamedTuple, Any, Any], None, None]:
+        evalStats = self.getEvalStats(predictedVarName)
+        for i, namedTuple in enumerate(self.inputData.itertuples()):
+            yield namedTuple, evalStats.y_predicted[i], evalStats.y_true[i]
+
+
+class VectorModelEvalStatsCollection(EvalStatsCollection[TVectorModelEvalStats]):
+    pass
