+++ src/sensai/hyperopt.py
@@ -8,13 +8,14 @@ from typing import Dict, Sequence, Any, Callable, Generator, Union, Tuple, List,
 
 import pandas as pd
 
+from .evaluation.util import computeEvaluationMetricsDict
+from .evaluation.vector_model import VectorModelEvaluator, VectorModelCrossValidator
 from .local_search import SACostValue, SACostValueNumeric, SAOperator, SAState, SimulatedAnnealing, \
     SAProbabilitySchedule, SAProbabilityFunctionLinear
+from .models.vector_model import VectorModel
 from .tracking.tracking_base import TrackedExperimentDataProvider, TrackedExperiment
-from .vector_model import VectorModel
-from .evaluation import VectorModelEvaluator, VectorModelCrossValidator, computeEvaluationMetricsDict
 
-_log = logging.getLogger(__name__)
+log = logging.getLogger(__name__)
 
 
 def iterParamCombinations(hyperParamValues: Dict[str, Sequence[Any]]) -> Generator[Dict[str, Any], None, None]:
@@ -87,7 +88,7 @@ class ParameterCombinationEquivalenceClassValueCache(ABC):
     def _equivalenceClass(self, params: Dict[str, Any]) -> Hashable:
         """
         Computes a (hashable) equivalence class representation for the given parameter combination.
-        For instance, if all parameters have influence on the evaluation of a model and no two combinations would 
+        For instance, if all parameters have influence on the evaluation of a model and no two combinations would
         lead to equivalent results, this could simply return a tuple containing all parameter values (in a fixed order).
 
         :param params: the parameter combination
@@ -112,15 +113,17 @@ class ParametersMetricsCollection:
     """
     Utility class for holding and persisting evaluation results
     """
-    def __init__(self, csvPath=None, sortColumnName=None):
+    def __init__(self, csvPath=None, sortColumnName=None, ascending=True):
         """
         :param csvPath: path to save the data frame to upon every update
         :param sortColumnName: the column name by which to sort the data frame that is collected; if None, do not sort
+        :param ascending: whether to sort in ascending order; ignored if sortColumnName is None
         """
         self.sortColumnName = sortColumnName
         self.csvPath = csvPath
         self.df = None
         self.cols = None
+        self.ascending = ascending
         self._currentRow = 0
 
     def addValues(self, values: Dict[str, Any]):
@@ -137,7 +140,7 @@ class ParametersMetricsCollection:
             # check sort column and move it to the front
             if self.sortColumnName is not None:
                 if self.sortColumnName not in self.cols:
-                    _log.warning(f"Specified sort column '{self.sortColumnName}' not in list of columns: {self.cols}; sorting will not take place!")
+                    log.warning(f"Specified sort column '{self.sortColumnName}' not in list of columns: {self.cols}; sorting will not take place!")
                 else:
                     self.cols.remove(self.sortColumnName)
                     self.cols.insert(0, self.sortColumnName)
@@ -150,7 +153,7 @@ class ParametersMetricsCollection:
 
         # sort where applicable
         if self.sortColumnName is not None and self.sortColumnName in self.df.columns:
-            self.df.sort_values(self.sortColumnName, axis=0, inplace=True)
+            self.df.sort_values(self.sortColumnName, axis=0, inplace=True, ascending=self.ascending)
             self.df.reset_index(drop=True, inplace=True)
 
         self._saveCSV()
@@ -171,7 +174,7 @@ class GridSearch(TrackedExperimentDataProvider):
     Instances of this class can be used for evaluating models with different user-provided parametrizations
     over the same data and persisting the results
     """
-    _log = _log.getChild(__qualname__)
+    log = log.getChild(__qualname__)
 
     def __init__(self, modelFactory: Callable[..., VectorModel], parameterOptions: Union[Dict[str, Sequence[Any]], List[Dict[str, Sequence[Any]]]],
             numProcesses=1, csvResultsPath=None, parameterCombinationSkipDecider: ParameterCombinationSkipDecider = None):
@@ -203,7 +206,7 @@ class GridSearch(TrackedExperimentDataProvider):
             for options in parameterOptions.values():
                 n *= len(options)
             self.numCombinations += n
-        _log.info(f"Created GridSearch object for {self.numCombinations} parameter combinations")
+        log.info(f"Created GridSearch object for {self.numCombinations} parameter combinations")
 
         self._executor = None
         self._trackedExperiment = None
@@ -212,9 +215,9 @@ class GridSearch(TrackedExperimentDataProvider):
     def _evalParams(cls, modelFactory, evaluatorOrValidator, skipDecider: ParameterCombinationSkipDecider, **params) -> Optional[Dict[str, Any]]:
         if skipDecider is not None:
             if skipDecider.isSkipped(params):
-                cls._log.info(f"Parameter combination is skipped according to {skipDecider}: {params}")
+                cls.log.info(f"Parameter combination is skipped according to {skipDecider}: {params}")
                 return None
-        cls._log.info(f"Evaluating {params}")
+        cls.log.info(f"Evaluating {params}")
         model = modelFactory(**params)
         values = computeEvaluationMetricsDict(model, evaluatorOrValidator)
         values["str(model)"] = str(model)
@@ -226,7 +229,8 @@ class GridSearch(TrackedExperimentDataProvider):
     def setTrackedExperiment(self, trackedExperiment: TrackedExperiment):
         self._trackedExperiment = trackedExperiment
 
-    def run(self, evaluatorOrValidator: Union[VectorModelEvaluator, VectorModelCrossValidator], sortColumnName=None) -> pd.DataFrame:
+    def run(self, evaluatorOrValidator: Union[VectorModelEvaluator, VectorModelCrossValidator],
+            sortColumnName=None) -> pd.DataFrame:
         """
         Run the grid search. If csvResultsPath was provided in the constructor, each evaluation result will be saved
         to that file directly after being computed
@@ -245,7 +249,7 @@ class GridSearch(TrackedExperimentDataProvider):
             if loggingCallback is not None:
                 loggingCallback(values)
             paramsMetricsCollection.addValues(values)
-            _log.info(f"Updated grid search result:\n{paramsMetricsCollection.getDataFrame().to_string()}")
+            log.info(f"Updated grid search result:\n{paramsMetricsCollection.getDataFrame().to_string()}")
 
         if self.numProcesses == 1:
             for parameterOptions in self.parameterOptionsList:
@@ -265,7 +269,7 @@ class GridSearch(TrackedExperimentDataProvider):
 
 
 class SAHyperOpt(TrackedExperimentDataProvider):
-    _log = _log.getChild(__qualname__)
+    log = log.getChild(__qualname__)
 
     class State(SAState):
         def __init__(self, params, randomState: Random, results: Dict, computeMetric: Callable[[Dict[str, Any]], float]):
@@ -286,12 +290,13 @@ class SAHyperOpt(TrackedExperimentDataProvider):
     class ParameterChangeOperator(SAOperator):
         def __init__(self, state: 'SAHyperOpt.State'):
             super().__init__(state)
+            self.state = state  # only needed in order for the type hints to work out
 
         def applyStateChange(self, params):
             self.state.params.update(params)
 
         def costDelta(self, params) -> SACostValue:
-            state: SAHyperOpt.State = self.state
+            state = self.state
             modelParams = dict(state.params)
             modelParams.update(params)
             return SACostValueNumeric(state.computeMetric(modelParams) - state.cost.value())
@@ -358,13 +363,13 @@ class SAHyperOpt(TrackedExperimentDataProvider):
         if parameterCombinationEquivalenceClassValueCache is not None:
             metrics = parameterCombinationEquivalenceClassValueCache.get(params)
         if metrics is not None:
-            cls._log.info(f"Result for parameter combination {params} could be retrieved from cache, not adding new result")
+            cls.log.info(f"Result for parameter combination {params} could be retrieved from cache, not adding new result")
             return metrics
         else:
-            cls._log.info(f"Evaluating parameter combination {params}")
+            cls.log.info(f"Evaluating parameter combination {params}")
             model = modelFactory(**params)
             metrics = computeEvaluationMetricsDict(model, evaluatorOrValidator)
-            cls._log.info(f"Got metrics {metrics} for {params}")
+            cls.log.info(f"Got metrics {metrics} for {params}")
 
             values = dict(metrics)
             values["str(model)"] = str(model)
@@ -373,7 +378,7 @@ class SAHyperOpt(TrackedExperimentDataProvider):
                 trackedExperiment.trackValues(values)
             if parametersMetricsCollection is not None:
                 parametersMetricsCollection.addValues(values)
-                cls._log.info(f"Data frame with all results:\n\n{parametersMetricsCollection.getDataFrame().to_string()}\n")
+                cls.log.info(f"Data frame with all results:\n\n{parametersMetricsCollection.getDataFrame().to_string()}\n")
             if parameterCombinationEquivalenceClassValueCache is not None:
                 parameterCombinationEquivalenceClassValueCache.set(params, metrics)
         return metrics
