+++ src/sensai/sklearn/sklearn_regression.py
@@ -6,10 +6,10 @@ import sklearn.neighbors
 import sklearn.neural_network
 import sklearn.svm
 
-from .sklearn_base import AbstractSkLearnMultipleOneDimVectorRegressionModel, AbstractSkLearnMultiDimVectorRegressionModel
+from .sklearn_base import AbstractSkLearnMultipleOneDimVectorRegressionModel, \
+    AbstractSkLearnMultiDimVectorRegressionModel
 
-
-_log = logging.getLogger(__name__)
+log = logging.getLogger(__name__)
 
 
 class SkLearnRandomForestVectorRegressionModel(AbstractSkLearnMultipleOneDimVectorRegressionModel):
